Move network files into network module.
