+++ src/adapters/template.rs
@@ -4,7 +4,7 @@ use crate::adapter::{
     Resource, Remote, Local, Adapter, SendStatus, AcceptedType, ReadStatus, ConnectionInfo,
     ListeningInfo,
 };
-use crate::remote_addr::{RemoteAddr};
+use crate::network::{RemoteAddr};
 
 use mio::event::{Source};
