+++ src/adapters/web_socket.rs
@@ -2,7 +2,7 @@ use crate::adapter::{
     Resource, Remote, Local, Adapter, SendStatus, AcceptedType, ReadStatus, ConnectionInfo,
     ListeningInfo,
 };
-use crate::remote_addr::{RemoteAddr};
+use crate::network::{RemoteAddr};
 use crate::util::thread::{OTHER_THREAD_ERR};
 
 use mio::event::{Source};
