+++ src/lib.rs
@@ -1,20 +1,16 @@
 //! See the [Github README](https://github.com/lemunozm/message-io),
 //! to see an overview of this library.
 
-// Tells rustdoc where is the README to compile and test the rust code there
 #[cfg(doctest)]
+// Tells rustdoc where is the README to compile and test the rust code there
 doc_comment::doctest!("../README.md");
 
-mod resource_id;
-mod endpoint;
-mod poll;
-mod registry;
-mod driver;
-mod engine;
 mod adapters;
-mod remote_addr;
-mod transport;
-mod network_thread;
+
+/// Module that specify the pattern to follow to create adapters.
+/// This module is not part of the public API itself,
+/// it must be used from the internals to build new adapters.
+pub mod adapter;
 
 /// Main module of `message-io`.
 /// It contains all the resources and tools to create and manage connections.
@@ -25,11 +21,6 @@ pub mod network;
 /// It can be used alone or along with the network module to reach a synchronized way to deal with
 /// events comming from the network.
 pub mod events;
-
-/// Module that specify the pattern to follow to create adapters.
-/// This module is not part of the public API itself,
-/// it must be used from the internals to build new adapters.
-pub mod adapter;
 
 /// General purpose utilities.
 pub mod util;
