+++ src/network/endpoint.rs
@@ -1,4 +1,4 @@
-use crate::resource_id::{ResourceId};
+use super::resource_id::{ResourceId};
 
 use std::net::{SocketAddr};
