+++ src/network/resource_id.rs
