+++ src/network/transport.rs
@@ -1,4 +1,4 @@
-use crate::engine::{AdapterLauncher};
+use super::engine::{AdapterLauncher};
 
 #[cfg(feature = "tcp")]
 use crate::adapters::tcp::{self, TcpAdapter};
